Move all models to legacy
