+++ src/app/models/legacy/category-filter.interface.ts
