+++ src/app/models/legacy/category-structure.interface.ts
