+++ src/app/models/legacy/image-type.enum.ts
