+++ src/app/models/legacy/photo.interface.ts
